refactor: Consolidate formatDuration, formatTimestamp, and formatElapsed
