+++ apps/ui/src/components/views/dashboard-view/event-feed.tsx
@@ -8,6 +8,7 @@ import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@prot
 import * as LucideIcons from 'lucide-react';
 import type { LucideIcon } from 'lucide-react';
 import { cn } from '@/lib/utils';
+import { formatTimestamp } from '@protolabsai/utils';
 
 interface EventFeedProps {
   projectPath: string | null;
@@ -22,25 +23,6 @@ function getIconComponent(iconName: string): LucideIcon {
     return (LucideIcons as unknown as Record<string, LucideIcon>)[iconName];
   }
   return LucideIcons.Circle;
-}
-
-/**
- * Format a timestamp for display
- */
-function formatTimestamp(timestamp: string): string {
-  const date = new Date(timestamp);
-  const now = new Date();
-  const diffMs = now.getTime() - date.getTime();
-  const diffMins = Math.floor(diffMs / 60000);
-  const diffHours = Math.floor(diffMs / 3600000);
-  const diffDays = Math.floor(diffMs / 86400000);
-
-  if (diffMins < 1) return 'Just now';
-  if (diffMins < 60) return `${diffMins}m ago`;
-  if (diffHours < 24) return `${diffHours}h ago`;
-  if (diffDays < 7) return `${diffDays}d ago`;
-
-  return date.toLocaleDateString();
 }
 
 /**
