+++ libs/ui/package.json
@@ -71,6 +71,7 @@
   },
   "dependencies": {
     "@protolabsai/types": "*",
+    "@protolabsai/utils": "*",
     "@radix-ui/react-checkbox": "^1.1.2",
     "@radix-ui/react-collapsible": "1.1.12",
     "@radix-ui/react-dialog": "^1.1.4",
