+++ libs/ui/src/ai/tool-results/agent-status-card.tsx
@@ -8,6 +8,7 @@
  */
 
 import { Loader2, Bot, CheckCircle2, XCircle, Clock } from 'lucide-react';
+import { formatElapsed } from '@protolabsai/utils';
 import { cn } from '../../lib/utils.js';
 import type { ToolResultRendererProps } from '../tool-result-registry.js';
 
@@ -42,16 +43,6 @@ function extractData(output: unknown): AgentStatusData | null {
   }
   if ('agent' in o || 'agents' in o || 'featureId' in o) return o as AgentStatusData;
   return o as AgentStatusData;
-}
-
-/** Format elapsed milliseconds into a human-readable string */
-function formatElapsed(ms: number): string {
-  if (ms < 1000) return `${ms}ms`;
-  const s = Math.floor(ms / 1000);
-  if (s < 60) return `${s}s`;
-  const m = Math.floor(s / 60);
-  const rem = s % 60;
-  return rem > 0 ? `${m}m ${rem}s` : `${m}m`;
 }
 
 type AgentDisplayStatus = 'running' | 'stopped' | 'done' | 'error' | 'unknown';
